requirements.txt
requests
pandas
numpy
scikit-learn
sentence-transformers
bertopic
umap-learn
matplotlib
seaborn
plotly
